% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalesOrder.R
\docType{data}
\name{SalesOrder_all_columns_ui}
\alias{SalesOrder_all_columns_ui}
\title{Total DeliveryQty'}
\format{
An object of class \code{character} of length 8.
}
\usage{
SalesOrder_all_columns_ui
}
\description{
Total DeliveryQty'
}
\keyword{datasets}
